+++ cmd/agent/main.go
@@ -13,8 +13,8 @@ import (
 	"github.com/cheesydui-cloud/mieru/internal/config"
 )
 
-// set by -ldflags "-X main.Version=v0.5.32"
-var Version = "v0.5.32"
+// set by -ldflags "-X main.Version=v0.5.33"
+var Version = "v0.5.33"
 
 func main() {
 	showVersion := flag.Bool("version", false, "print version and exit")